examples/inference/stablediffusion-python-tensorflow/requirements.txt
setuptools>=69.1.0
keras_cv==0.6.4
pillow==10.2.0
tensorflow-cpu==2.13.1; python_version >= '3.9'
tensorflow==2.13.1; python_version < '3.9'
